Fix data type
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@120 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/LuceneKit/Source/Index/LCSegmentReader.m
@@ -334,7 +334,7 @@
   return [[LCSegmentTermPositions alloc] initWithSegmentReader: self];
 }
 
-- (int) docFreq: (LCTerm *) t
+- (long) docFreq: (LCTerm *) t
 {
     LCTermInfo *ti = [tis termInfo: t];
     if (ti != nil)
